Add conditions to modify position and styles of QueuedSongList at breakpoint
Using the useMediaQuery hook, the player will be positioned to the bottom of the screen (fixed position) until the screen is > the medium breakpoint. At that point, it will be positioned to the right of the screen. The QueuedList will only appear if the screen size is > than the medium breakpoint.